+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IssuingStockRule resource
 - Rules and DebtorWorkspaceID attributes to CreditNote resource
 - PixFraud resource
+- PixKeyHolmes resource
 
 ## [0.12.0] - 2026-02-10
 ### Fixed